SQL-1991: remove all enums from odbc api parameters (#195)
* make desc u16

* add tests for col attr

* remve nullability parameters

* remove all enums from func defs

* add updated definitions

* fix power bi tests

* fix more integration tests

* fix failures

* add invalid driver completion error

* add driver not capable to sqlcolattribute

* finalize new errors

* setdescfield param update

* clarify invalid vs unsupported col attr

+++ integration_test/src/test_runner/mod.rs
@@ -183,7 +183,7 @@ pub fn run_resultset_tests_odbc_2(generate: bool) -> Result<()> {
             SqlReturn::SUCCESS,
             definitions::SQLSetEnvAttr(
                 env,
-                EnvironmentAttribute::SQL_ATTR_ODBC_VERSION,
+                EnvironmentAttribute::SQL_ATTR_ODBC_VERSION as i32,
                 2 as *mut _,
                 0,
             )
@@ -331,7 +331,10 @@ fn run_function_test(
             unsafe {
                 let data_type: definitions::SqlDataType =
                     std::mem::transmute(function[1].as_i64().unwrap() as i16);
-                Ok(definitions::SQLGetTypeInfo(statement as HStmt, data_type))
+                Ok(definitions::SQLGetTypeInfo(
+                    statement as HStmt,
+                    data_type as i16,
+                ))
             }
         }
         "sqltables" => {
@@ -694,7 +697,7 @@ fn get_column_attribute(
         match definitions::SQLColAttributeW(
             stmt as *mut _,
             column as USmallInt,
-            field_identifier,
+            field_identifier as u16,
             character_attrib_ptr,
             BUFFER_LENGTH as SmallInt,
             string_length_ptr,
@@ -730,7 +733,7 @@ fn get_data(stmt: HStmt, column: USmallInt, data_type: CDataType) -> Result<Valu
             stmt as *mut _,
             // Result set columns start at 1, the column input parameter is 0-indexed
             column + 1,
-            data_type,
+            data_type as i16,
             buffer as *mut _,
             BUFFER_LENGTH as isize,
             out_len_or_ind,
